chore: 1098 calibrate detail pane (#1128)
