+++ packages/client/hmi-client/src/components/models/tera-model.vue
@@ -135,89 +135,11 @@
 		<template v-if="modelView === ModelView.MODEL">
 			<Accordion :multiple="true" :active-index="[0, 1, 2, 3, 4]">
 				<AccordionTab header="Model diagram">
-					<section class="model_diagram">
-						<TeraResizablePanel>
-							<div ref="splitterContainer" class="splitter-container">
-								<Splitter :gutterSize="5" :layout="layout">
-									<SplitterPanel
-										class="tera-split-panel"
-										:size="equationPanelSize"
-										:minSize="equationPanelMinSize"
-										:maxSize="equationPanelMaxSize"
-									>
-										<section class="graph-element">
-											<Toolbar>
-												<template #start>
-													<Button
-														@click="resetZoom"
-														label="Reset zoom"
-														class="p-button-sm p-button-outlined toolbar-button"
-													/>
-												</template>
-												<template #center>
-													<span class="toolbar-subgroup">
-														<Button
-															v-if="isEditing"
-															@click="addState"
-															label="Add state"
-															class="p-button-sm p-button-outlined toolbar-button"
-														/>
-														<Button
-															v-if="isEditing"
-															@click="addTransition"
-															label="Add transition"
-															class="p-button-sm p-button-outlined toolbar-button"
-														/>
-													</span>
-												</template>
-												<template #end>
-													<span class="toolbar-subgroup">
-														<Button
-															v-if="isEditing"
-															@click="cancelEdit"
-															label="Cancel"
-															class="p-button-sm p-button-outlined toolbar-button"
-														/>
-														<Button
-															@click="toggleEditMode"
-															:label="isEditing ? 'Save model' : 'Edit model'"
-															:class="
-																isEditing
-																	? 'p-button-sm toolbar-button-saveModel'
-																	: 'p-button-sm p-button-outlined toolbar-button'
-															"
-														/>
-													</span>
-												</template>
-											</Toolbar>
-											<div v-if="model" ref="graphElement" class="graph-element" />
-											<ContextMenu ref="menu" :model="contextMenuItems" />
-										</section>
-									</SplitterPanel>
-									<SplitterPanel
-										class="tera-split-panel"
-										:size="mathPanelSize"
-										:minSize="mathPanelMinSize"
-										:maxSize="mathPanelMaxSize"
-									>
-										<section class="math-editor-container" :class="mathEditorSelected">
-											<tera-math-editor
-												:is-editable="isEditable"
-												:latex-equation="equationLatex"
-												:is-editing-eq="isEditingEQ"
-												:is-math-ml-valid="isMathMLValid"
-												:math-mode="MathEditorModes.LIVE"
-												@cancel-editing="cancelEditng"
-												@equation-updated="setNewLatexFormula"
-												@validate-mathml="validateMathML"
-												@set-editing="isEditingEQ = true"
-											></tera-math-editor>
-										</section>
-									</SplitterPanel>
-								</Splitter>
-							</div>
-						</TeraResizablePanel>
-					</section>
+					<tera-model-diagram
+						:model="model"
+						:is-editable="props.isEditable"
+						@update-model-content="updateModelContent"
+					/>
 				</AccordionTab>
 				<AccordionTab v-if="model">
 					<template #header> Model configurations </template>
@@ -352,40 +274,16 @@
 </template>
 
 <script setup lang="ts">
-import { remove, isEmpty, pickBy, isArray, cloneDeep } from 'lodash';
-import { IGraph } from '@graph-scaffolder/index';
-import {
-	watch,
-	ref,
-	computed,
-	onMounted,
-	onUnmounted,
-	onUpdated,
-	PropType,
-	ComputedRef
-} from 'vue';
-import { runDagreLayout } from '@/services/graph';
-import { PetrinetRenderer } from '@/petrinet/petrinet-renderer';
-import {
-	parsePetriNet2IGraph,
-	PetriNet,
-	NodeData,
-	EdgeData,
-	parseIGraph2PetriNet,
-	mathmlToPetri,
-	petriToLatex,
-	NodeType
-} from '@/petrinet/petrinet-service';
+import { isEmpty, cloneDeep } from 'lodash';
+import { watch, ref, computed, onUpdated, PropType, ComputedRef } from 'vue';
+import { PetriNet, parseIGraph2PetriNet } from '@/petrinet/petrinet-service';
 import Textarea from 'primevue/textarea';
-import { bucky } from '@/temp/buckyAMR';
 import InputText from 'primevue/inputtext';
-import { separateEquations, MathEditorModes } from '@/utils/math';
-import { getModel, updateModel, createModel, addModelToProject } from '@/services/model';
-import { getRelatedArtifacts } from '@/services/provenance';
+import { bucky } from '@/temp/buckyAMR';
+import { createModel, addModelToProject, getModel } from '@/services/model';
 import { useRouter } from 'vue-router';
 import { RouteName } from '@/router/routes';
 import useResourcesStore from '@/stores/resources';
-import { logger } from '@/utils/logger';
 import Button from 'primevue/button';
 import Tag from 'primevue/tag';
 import Accordion from 'primevue/accordion';
@@ -396,25 +294,21 @@ import DataTable from 'primevue/datatable';
 import Column from 'primevue/column';
 import Row from 'primevue/row';
 import ColumnGroup from 'primevue/columngroup';
-import ContextMenu from 'primevue/contextmenu';
 import * as textUtil from '@/utils/text';
 import ForecastLauncher from '@/components/models/tera-forecast-launcher.vue';
 import { isModel, isDataset, isDocument } from '@/utils/data-util';
 import { ITypedModel, Model } from '@/types/Model';
 import { AskemModelRepresentationType } from '@/types/AskemModelRepresentation';
 import { ResultType } from '@/types/common';
-import { Document, ProvenanceType, Dataset } from '@/types/Types';
-import TeraMathEditor from '@/components/mathml/tera-math-editor.vue';
-import Splitter from 'primevue/splitter';
-import SplitterPanel from 'primevue/splitterpanel';
+import { Document, Dataset, ProvenanceType } from '@/types/Types';
 import TeraAsset from '@/components/asset/tera-asset.vue';
-import Toolbar from 'primevue/toolbar';
 import { FilterMatchMode } from 'primevue/api';
 import { IProject, ProjectAssetTypes } from '@/types/Project';
 import TeraModal from '@/components/widgets/tera-modal.vue';
 import { useOpenedWorkflowNodeStore } from '@/stores/opened-workflow-node';
 import TeraAssetNav from '@/components/asset/tera-asset-nav.vue';
-import TeraResizablePanel from '@/components/widgets/tera-resizable-panel.vue';
+import { getRelatedArtifacts } from '@/services/provenance';
+import TeraModelDiagram from './tera-model-diagram.vue';
 
 interface StringValueMap {
 	[key: string]: string;
@@ -472,7 +366,6 @@ const resources = useResourcesStore();
 const router = useRouter();
 
 const relatedTerariumArtifacts = ref<ResultType[]>([]);
-const menu = ref();
 
 const model = ref<ITypedModel<PetriNet> | null>(null);
 const amr = ref<AskemModelRepresentationType | null>(null);
@@ -484,31 +377,9 @@ const newModelName = ref('New Model');
 const newDescription = ref<string | undefined>('');
 const newPetri = ref();
 
-const selectedVariable = ref('');
-
-const equationLatex = ref<string>('');
-const equationLatexOriginal = ref<string>('');
-const equationLatexNew = ref<string>('');
 const isMathMLValid = ref<boolean>(true);
 
-const splitterContainer = ref<HTMLElement | null>(null);
-const layout = ref<'horizontal' | 'vertical' | undefined>('horizontal');
 const showForecastLauncher = ref(false);
-
-const switchWidthPercent = ref<number>(50); // switch model layout when the size of the model window is < 50%
-
-const equationPanelSize = ref<number>(50);
-const equationPanelMinSize = ref<number>(0);
-const equationPanelMaxSize = ref<number>(100);
-
-const mathPanelSize = ref<number>(50);
-const mathPanelMinSize = ref<number>(0);
-const mathPanelMaxSize = ref<number>(100);
-
-const graphElement = ref<HTMLDivElement | null>(null);
-let renderer: PetrinetRenderer | null = null;
-let eventX = 0;
-let eventY = 0;
 
 const modelConfigNames = ref([{ name: 'Config 1' }]);
 const fakeExtractions = ref(['Resource 1', 'Resource 2', 'Resource 3']);
@@ -595,6 +466,10 @@ const onCellEditComplete = (event) => {
 	}
 };
 
+function updateModelContent(rendererGraph) {
+	if (model.value) model.value.content = parseIGraph2PetriNet(rendererGraph);
+}
+
 const onCellEditStart = (event) => {
 	const { data, field, index } = event;
 
@@ -666,44 +541,6 @@ watch(
 	}
 );
 
-const updateLayout = () => {
-	if (splitterContainer.value) {
-		layout.value =
-			(splitterContainer.value.offsetWidth / window.innerWidth) * 100 < switchWidthPercent.value ||
-			window.innerWidth < 800
-				? 'vertical'
-				: 'horizontal';
-	}
-};
-
-const handleResize = () => {
-	updateLayout();
-};
-
-onMounted(() => {
-	window.addEventListener('resize', handleResize);
-	handleResize();
-	// new model
-	if (props.assetId === '') {
-		isEditingEQ.value = true;
-		isMathMLValid.value = false;
-	}
-});
-
-onUnmounted(() => {
-	window.removeEventListener('resize', handleResize);
-});
-
-const mathEditorSelected = computed(() => {
-	if (!isMathMLValid.value) {
-		return 'math-editor-error';
-	}
-	if (isEditingEQ.value) {
-		return 'math-editor-selected';
-	}
-	return '';
-});
-
 const globalFilter = ref({
 	// @ts-ignore
 	// eslint-disable-line
@@ -722,39 +559,6 @@ const globalFilter = ref({
 // 	}
 // };
 
-const onVariableSelected = (variable: string) => {
-	if (variable) {
-		if (variable === selectedVariable.value) {
-			selectedVariable.value = '';
-			equationLatex.value = equationLatexOriginal.value;
-		} else {
-			selectedVariable.value = variable;
-			equationLatex.value = equationLatexOriginal.value.replaceAll(
-				selectedVariable.value,
-				String.raw`{\color{red}${variable}}`
-			);
-		}
-		renderer?.toggoleNodeSelectionByLabel(variable);
-	} else {
-		equationLatex.value = equationLatexOriginal.value;
-	}
-};
-
-const setNewLatexFormula = (formulaString: string) => {
-	equationLatexNew.value = formulaString;
-};
-
-const updateLatexFormula = (formulaString: string) => {
-	equationLatex.value = formulaString;
-	equationLatexOriginal.value = formulaString;
-};
-
-const cancelEditng = () => {
-	isEditingEQ.value = false;
-	isMathMLValid.value = true;
-	updateLatexFormula(equationLatexOriginal.value);
-};
-
 const relatedTerariumModels = computed(
 	() => relatedTerariumArtifacts.value.filter((d) => isModel(d)) as Model[]
 );
@@ -765,6 +569,13 @@ const relatedTerariumDocuments = computed(
 	() => relatedTerariumArtifacts.value.filter((d) => isDocument(d)) as Document[]
 );
 
+// Highlight strings based on props.highlight
+function highlightSearchTerms(text: string | undefined): string {
+	if (!!props.highlight && !!text) {
+		return textUtil.highlight(text, props.highlight);
+	}
+	return text ?? '';
+}
 const fetchRelatedTerariumArtifacts = async () => {
 	if (model.value) {
 		const results = await getRelatedArtifacts(props.assetId, ProvenanceType.ModelRevision);
@@ -774,35 +585,15 @@ const fetchRelatedTerariumArtifacts = async () => {
 	}
 };
 
-// Highlight strings based on props.highlight
-function highlightSearchTerms(text: string | undefined): string {
-	if (!!props.highlight && !!text) {
-		return textUtil.highlight(text, props.highlight);
-	}
-	return text ?? '';
-}
-
-// Whenever selectedModelId changes, fetch model with that ID
 watch(
 	() => [props.assetId],
 	async () => {
-		updateLatexFormula('');
-		amr.value = null;
 		if (props.assetId !== '') {
-			const result = await getModel(props.assetId);
-			if (result && result.name === 'Bucky') {
+			model.value = await getModel(props.assetId);
+			if (model.value && model.value.name === 'Bucky') {
 				amr.value = bucky;
-				model.value = result;
-			} else {
-				model.value = result;
 			}
 			fetchRelatedTerariumArtifacts();
-			if (model.value) {
-				const data = await petriToLatex(model.value.content);
-				if (data) {
-					updateLatexFormula(data);
-				}
-			}
 		} else {
 			model.value = null;
 		}
@@ -825,169 +616,8 @@ onUpdated(() => {
 	}
 });
 
-const editorKeyHandler = (event: KeyboardEvent) => {
-	// Ignore backspace if the current focus is a text/input box
-	if ((event.target as HTMLElement).tagName === 'INPUT') {
-		return;
-	}
-
-	if (event.key === 'Backspace' && renderer) {
-		if (renderer && renderer.nodeSelection) {
-			const nodeData = renderer.nodeSelection.datum();
-			remove(renderer.graph.edges, (e) => e.source === nodeData.id || e.target === nodeData.id);
-			remove(renderer.graph.nodes, (n) => n.id === nodeData.id);
-			renderer.nodeSelection = null;
-			renderer.render();
-		}
-
-		if (renderer && renderer.edgeSelection) {
-			const edgeData = renderer.edgeSelection.datum();
-			remove(
-				renderer.graph.edges,
-				(e) => e.source === edgeData.source && e.target === edgeData.target
-			);
-			renderer.edgeSelection = null;
-			renderer.render();
-		}
-	}
-	if (event.key === 'Enter' && renderer) {
-		if (renderer.nodeSelection) {
-			renderer.deselectNode(renderer.nodeSelection);
-			renderer.nodeSelection
-				.selectAll('.no-drag')
-				.style('opacity', 0)
-				.style('visibility', 'hidden');
-			renderer.nodeSelection = null;
-		}
-		if (renderer.edgeSelection) {
-			renderer.deselectEdge(renderer.edgeSelection);
-			renderer.edgeSelection = null;
-		}
-	}
-};
-
-// Model editor context menu
-const contextMenuItems = ref([
-	{
-		label: 'Add state',
-		icon: 'pi pi-fw pi-circle',
-		command: () => {
-			if (renderer) {
-				renderer.addNode('S', '?', { x: eventX, y: eventY });
-			}
-		}
-	},
-	{
-		label: 'Add transition',
-		icon: 'pi pi-fw pi-stop',
-		command: () => {
-			if (renderer) {
-				renderer.addNode('T', '?', { x: eventX, y: eventY });
-			}
-		}
-	}
-]);
-
-// Render graph whenever a new model is fetched or whenever the HTML element
-//	that we render the graph to changes.
-watch(
-	[model, graphElement],
-	async () => {
-		if (model.value === null || graphElement.value === null) return;
-		// Convert petri net into a graph
-		const graphData: IGraph<NodeData, EdgeData> = parsePetriNet2IGraph(model.value.content, {
-			S: { width: 60, height: 60 },
-			T: { width: 40, height: 40 }
-		});
-
-		// Create renderer
-		renderer = new PetrinetRenderer({
-			el: graphElement.value as HTMLDivElement,
-			useAStarRouting: false,
-			useStableZoomPan: true,
-			runLayout: runDagreLayout,
-			dragSelector: 'no-drag'
-		});
-
-		renderer.on('add-edge', (_evtName, _evt, _selection, d) => {
-			renderer?.addEdge(d.source, d.target);
-		});
-
-		renderer.on('background-contextmenu', (_evtName, evt, _selection, _renderer, pos: any) => {
-			if (!renderer?.editMode) return;
-			eventX = pos.x;
-			eventY = pos.y;
-			menu.value.show(evt);
-		});
-
-		renderer.on('background-click', () => {
-			if (menu.value) menu.value.hide();
-
-			// de-select node if selection exists
-			if (selectedVariable.value) {
-				onVariableSelected(selectedVariable.value);
-			}
-		});
-
-		renderer.on('node-click', async (_evtName, _evt, _e, _renderer, d) => {
-			// Note: do not change the renderer's visuals, this is done internally
-			onVariableSelected(d.label);
-		});
-
-		// Render graph
-		await renderer?.setData(graphData);
-		await renderer?.render();
-		const latexFormula = await petriToLatex(model.value.content);
-		if (latexFormula) {
-			updateLatexFormula(latexFormula);
-		} else {
-			updateLatexFormula('');
-		}
-	},
-	{ deep: true }
-);
-
-const updatePetri = async (m: PetriNet) => {
-	// equationML.value = mathmlString;
-	// Convert petri net into a graph
-	const graphData: IGraph<NodeData, EdgeData> = parsePetriNet2IGraph(m, {
-		S: { width: 60, height: 60 },
-		T: { width: 40, height: 40 }
-	});
-
-	// Create renderer
-	renderer = new PetrinetRenderer({
-		el: graphElement.value as HTMLDivElement,
-		useAStarRouting: false,
-		useStableZoomPan: true,
-		runLayout: runDagreLayout,
-		dragSelector: 'no-drag'
-	});
-
-	renderer.on('add-edge', (_evtName, _evt, _selection, d) => {
-		renderer?.addEdge(d.source, d.target);
-	});
-
-	renderer.on('background-contextmenu', (_evtName, evt, _selection, _renderer, pos: any) => {
-		if (!renderer?.editMode) return;
-		eventX = pos.x;
-		eventY = pos.y;
-		menu.value.toggle(evt);
-	});
-
-	// Render graph
-	await renderer?.setData(graphData);
-	await renderer?.render();
-	updateLatexFormula(equationLatexNew.value);
-};
-
 const launchForecast = () => {
 	showForecastLauncher.value = true;
-};
-
-const hasNoEmptyKeys = (obj: Record<string, unknown>): boolean => {
-	const nonEmptyKeysObj = pickBy(obj, (value) => !isEmpty(value));
-	return Object.keys(nonEmptyKeysObj).length === Object.keys(obj).length;
 };
 
 const createNewModel = async () => {
@@ -1019,38 +649,6 @@ const createNewModel = async () => {
 	}
 };
 
-const validateMathML = async (mathMlString: string, editMode: boolean) => {
-	isEditingEQ.value = true;
-	isMathMLValid.value = false;
-	const cleanedMathML = separateEquations(mathMlString);
-	if (mathMlString === '') {
-		isMathMLValid.value = true;
-		isEditingEQ.value = false;
-	} else if (!editMode) {
-		try {
-			newPetri.value = await mathmlToPetri(cleanedMathML);
-			if (
-				(isArray(newPetri.value) && newPetri.value.length > 0) ||
-				(!isArray(newPetri.value) &&
-					Object.keys(newPetri.value).length > 0 &&
-					hasNoEmptyKeys(newPetri.value))
-			) {
-				isMathMLValid.value = true;
-				isEditingEQ.value = false;
-				updatePetri(newPetri.value);
-			} else {
-				logger.error(
-					'MathML cannot be converted to a Petrinet.  Please try again or click cancel.'
-				);
-			}
-		} catch (e) {
-			isMathMLValid.value = false;
-		}
-	} else if (editMode) {
-		isMathMLValid.value = true;
-	}
-};
-
 const goToSimulationRunPage = () => {
 	showForecastLauncher.value = false;
 	router.push({
@@ -1061,59 +659,6 @@ const goToSimulationRunPage = () => {
 			pageType: 'simulation_runs'
 		}
 	});
-};
-
-onMounted(async () => {
-	fetchRelatedTerariumArtifacts();
-	document.addEventListener('keyup', editorKeyHandler);
-});
-
-onUnmounted(() => {
-	document.removeEventListener('keyup', editorKeyHandler);
-});
-
-const toggleEditMode = () => {
-	isEditing.value = !isEditing.value;
-	renderer?.setEditMode(isEditing.value);
-	if (!isEditing.value && model.value && renderer) {
-		model.value.content = parseIGraph2PetriNet(renderer.graph);
-		updateModel(model.value);
-	} else if (isEditing.value && selectedVariable.value) {
-		// de-select node if selection exists
-		onVariableSelected(selectedVariable.value);
-	}
-};
-
-// Cancel existing edits, currently this will:
-// - Resets changs to the model structure
-const cancelEdit = async () => {
-	isEditing.value = false;
-	if (!model.value) return;
-
-	// Convert petri net into a graph with raw input data
-	const graphData: IGraph<NodeData, EdgeData> = parsePetriNet2IGraph(model.value.content, {
-		S: { width: 60, height: 60 },
-		T: { width: 40, height: 40 }
-	});
-
-	if (renderer) {
-		renderer.setEditMode(false);
-		await renderer.setData(graphData);
-		renderer.isGraphDirty = true;
-		await renderer.render();
-	}
-};
-
-const resetZoom = async () => {
-	renderer?.setToDefaultZoom();
-};
-
-const addState = async () => {
-	renderer?.addNodeCenter(NodeType.State, '?');
-};
-
-const addTransition = async () => {
-	renderer?.addNodeCenter(NodeType.Transition, '?');
 };
 
 const name = computed(() => highlightSearchTerms(model.value?.name ?? ''));
