+++ packages/client/hmi-client/src/components/workflow/simulate-operation.ts
@@ -1,7 +1,7 @@
-import { Operation } from '@/types/workflow';
+import { Operation, WorkflowOperationTypes } from '@/types/workflow';
 
 export const SimulateOperation: Operation = {
-	name: 'SimulateOperation',
+	name: WorkflowOperationTypes.SIMULATE,
 	description: 'given a model id, and configuration id, run a simulation',
 	inputs: [{ type: 'modelConfig' }],
 	outputs: [],
