+++ packages/client/hmi-client/src/components/workflow/tera-calibration.vue
@@ -1,39 +1,41 @@
 <template>
-	<Dropdown
-		placeholder="Timestep Column Name"
-		class="p-button dropdown-button"
-		:options="datasetColumnNames"
-		v-model="timestepColumnName"
-	/>
-	<table v-if="featureMap">
-		<tr>
-			<th>Dataset Column Name</th>
-			<th>Model Column Name</th>
-		</tr>
-		<tr v-for="(content, index) in featureMap" :key="index">
-			<Dropdown
-				placeholder="Dataset Column Name"
-				class="p-button dropdown-button"
-				:options="datasetColumnNames"
-				v-model="featureMap[index][0]"
-			/>
-			<td>{{ content[1] }}</td>
-		</tr>
-	</table>
-	<Button @click="startCalibration">Start Calibration Job</Button>
-	<form>
-		<label for="calibrationStatus">
-			<input v-model="runId" type="text" placeholder="Run ID" />
-		</label>
-		<Button @click="getCalibrationStatus"> Get Run Status </Button>
-	</form>
+	<tera-asset :name="'CALIBRATION NAME HERE'" :is-editable="false">
+		<Dropdown
+			placeholder="Timestep Column Name"
+			class="p-button dropdown-button"
+			:options="datasetColumnNames"
+			v-model="timestepColumnName"
+		/>
+		<table v-if="featureMap">
+			<tr>
+				<th>Dataset Column Name</th>
+				<th>Model Column Name</th>
+			</tr>
+			<tr v-for="(content, index) in featureMap" :key="index">
+				<Dropdown
+					placeholder="Dataset Column Name"
+					class="p-button dropdown-button"
+					:options="datasetColumnNames"
+					v-model="featureMap[index][0]"
+				/>
+				<td>{{ content[1] }}</td>
+			</tr>
+		</table>
+		<Button @click="startCalibration">Start Calibration Job</Button>
+		<form>
+			<label for="calibrationStatus">
+				<input v-model="runId" type="text" placeholder="Run ID" />
+			</label>
+			<Button @click="getCalibrationStatus"> Get Run Status </Button>
+		</form>
 
-	<form>
-		<label for="calibrationResult">
-			<input v-model="runId" type="text" placeholder="Run ID" />
-		</label>
-		<Button @click="getCalibrationResults"> Get Run Results </Button>
-	</form>
+		<form>
+			<label for="calibrationResult">
+				<input v-model="runId" type="text" placeholder="Run ID" />
+			</label>
+			<Button @click="getCalibrationResults"> Get Run Results </Button>
+		</form>
+	</tera-asset>
 </template>
 
 <script setup lang="ts">
@@ -44,9 +46,9 @@ import { CalibrationParams, CsvAsset } from '@/types/Types';
 import { ModelConfig } from '@/types/ModelConfig';
 import Dropdown from 'primevue/dropdown';
 import { downloadRawFile } from '@/services/dataset';
-import { WorkflowNode, WorkflowPort, WorkflowPortStatus } from '@/types/workflow';
-import { shimPetriModel } from '@/services/models/petri-shim';
-// import { calibrationParamExample } from '@/temp/calibrationExample';
+import { PetriNet } from '@/petrinet/petrinet-service';
+import { WorkflowNode } from '@/types/workflow';
+import TeraAsset from '@/components/asset/tera-asset.vue';
 
 const props = defineProps<{
 	node: WorkflowNode;
@@ -54,7 +56,7 @@ const props = defineProps<{
 const modelConfig = computed(() => props.node.inputs[0].value as ModelConfig | undefined);
 const datasetId = computed(() => props.node.inputs[1].value as number | undefined);
 
-const runId = ref('');
+const runId = ref(props.node.outputs[0].value as number | undefined);
 const timestepColumnName = ref<string>('');
 const datasetColumnNames = ref<string[]>();
 const modelColumnNames = computed(() =>
@@ -62,13 +64,24 @@ const modelColumnNames = computed(() =>
 );
 const csvAsset = shallowRef<CsvAsset | undefined>(undefined);
 const datasetValue = ref();
-const featureMap = ref();
-
-const emit = defineEmits(['append-output-port']);
+const featureMap = computed(() => modelColumnNames.value.map((stateName) => ['', stateName]));
 
 const startCalibration = async () => {
 	// Make calibration job.
+	// FIXME: current need to strip out metadata, should do serverside
+	const cleanedModel: PetriNet = {
+		S: [],
+		T: [],
+		I: [],
+		O: []
+	};
 	if (modelConfig.value) {
+		// Take out all the extra content in model.content
+		cleanedModel.S = modelConfig.value.model.content.S.map((s) => ({ sname: s.sname }));
+		cleanedModel.T = modelConfig.value.model.content.T.map((t) => ({ tname: t.tname }));
+		cleanedModel.I = modelConfig.value.model.content.I;
+		cleanedModel.O = modelConfig.value.model.content.O;
+
 		if (featureMap.value) {
 			const featureObject: { [index: string]: string } = {};
 			// Go from 2D array to a index: value like they want
@@ -78,7 +91,7 @@ const startCalibration = async () => {
 			}
 
 			const calibrationParam: CalibrationParams = {
-				model: shimPetriModel(modelConfig.value.model),
+				model: JSON.stringify(cleanedModel),
 				initials: modelConfig.value.initialValues,
 				params: modelConfig.value.parameterValues,
 				timesteps_column: timestepColumnName.value,
@@ -88,22 +101,12 @@ const startCalibration = async () => {
 			console.log(calibrationParam);
 			const results = await makeCalibrateJob(calibrationParam);
 			runId.value = results.id;
-			const outputPort: WorkflowPort = {
-				id: '1',
-				type: 'String',
-				status: WorkflowPortStatus.CONNECTED,
-				value: runId.value
-			};
-			emit('append-output-port', {
-				type: 'string',
-				label: 'Calibration Job ID',
-				value: { outputPort }
-			});
 		}
 	}
 };
 
 const getCalibrationStatus = async () => {
+	console.log('Getting status of run');
 	const results = await getRunStatus(Number(runId.value));
 	console.log('Done');
 	console.log(results);
@@ -115,16 +118,6 @@ const getCalibrationResults = async () => {
 	console.log('Done');
 	console.log(results);
 };
-
-watch(
-	() => modelConfig.value,
-	async () => {
-		if (modelConfig.value) {
-			// initialize featureMap
-			featureMap.value = modelColumnNames.value.map((stateName) => ['', stateName]);
-		}
-	}
-);
 
 watch(
 	() => datasetId.value, // When dataset ID changes, update datasetColumnNames
@@ -149,6 +142,7 @@ watch(
 :deep(.p-dropdown .p-dropdown-label.p-inputtext) {
 	color: white;
 }
+
 :deep(.p-inputtext) {
 	color: white;
 }
