+++ packages/client/hmi-client/src/components/workflow/tera-workflow-node.vue
@@ -143,9 +143,7 @@ function showNodeDrilldown() {
 			default:
 				break;
 		}
-		if (pageType && assetId) {
-			openedWorkflowNodeStore.setDrilldown(assetId, pageType);
-		}
+		openedWorkflowNodeStore.setDrilldown(assetId, pageType, props.node);
 	} else logger.error('Node needs a valid output', { silent: true });
 }
