+++ packages/client/hmi-client/src/stores/opened-workflow-node.ts
@@ -1,5 +1,6 @@
 import { defineStore } from 'pinia';
 import { ProjectAssetTypes } from '@/types/Project';
+import { WorkflowNode } from '@/types/workflow';
 
 interface StringValueMap {
 	[key: string]: number;
@@ -13,12 +14,18 @@ export const useOpenedWorkflowNodeStore = defineStore('opened-workflow-node', {
 		selectedOutputIndex: 0 as number,
 		// model node
 		initialValues: null as StringValueMap[] | null,
-		parameterValues: null as StringValueMap[] | null
+		parameterValues: null as StringValueMap[] | null,
+		node: null as WorkflowNode | null
 	}),
 	actions: {
-		setDrilldown(assetId: string | null, pageType: ProjectAssetTypes | null) {
+		setDrilldown(
+			assetId: string | null,
+			pageType: ProjectAssetTypes | null,
+			node: WorkflowNode | null
+		) {
 			this.assetId = assetId;
 			this.pageType = pageType;
+			this.node = node;
 		},
 		setModelConfig(initialValues: StringValueMap[], parameterValues: StringValueMap[]) {
 			this.initialValues = initialValues;
